CodeWhisperer: Add telemetry optout status to SendTelemetryEventRequest
Updated the API and put the field in SendTelemetryEvent API requests.
