+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/codewhisperer/credentials/CodeWhispererClientAdaptor.kt
@@ -21,6 +21,7 @@ import software.amazon.awssdk.services.codewhispererruntime.model.CreateUploadUr
 import software.amazon.awssdk.services.codewhispererruntime.model.CreateUploadUrlResponse
 import software.amazon.awssdk.services.codewhispererruntime.model.GenerateCompletionsRequest
 import software.amazon.awssdk.services.codewhispererruntime.model.GenerateCompletionsResponse
+import software.amazon.awssdk.services.codewhispererruntime.model.OptOutPreference
 import software.amazon.awssdk.services.codewhispererruntime.model.SendTelemetryEventResponse
 import software.amazon.awssdk.services.codewhispererruntime.model.SuggestionState
 import software.aws.toolkits.core.utils.getLogger
@@ -39,6 +40,7 @@ import software.aws.toolkits.jetbrains.services.codewhisperer.service.RequestCon
 import software.aws.toolkits.jetbrains.services.codewhisperer.service.ResponseContext
 import software.aws.toolkits.jetbrains.services.codewhisperer.util.CodeWhispererConstants
 import software.aws.toolkits.jetbrains.services.codewhisperer.util.transform
+import software.aws.toolkits.jetbrains.settings.AwsSettings
 import software.aws.toolkits.telemetry.CodewhispererCompletionType
 import software.aws.toolkits.telemetry.CodewhispererSuggestionState
 import java.time.Instant
@@ -211,6 +213,7 @@ open class CodeWhispererClientAdaptorImpl(override val project: Project) : CodeW
                     it.generatedLine(lineCount)
                 }
             }
+            requestBuilder.optOutPreference(getTelemetryOptoutPreference())
         }
     }
 
@@ -227,6 +230,7 @@ open class CodeWhispererClientAdaptorImpl(override val project: Project) : CodeW
                 it.timestamp(Instant.now())
             }
         }
+        requestBuilder.optOutPreference(getTelemetryOptoutPreference())
     }
 
     override fun sendUserModificationTelemetry(
@@ -246,6 +250,7 @@ open class CodeWhispererClientAdaptorImpl(override val project: Project) : CodeW
                 it.timestamp(Instant.now())
             }
         }
+        requestBuilder.optOutPreference(getTelemetryOptoutPreference())
     }
 
     override fun sendCodeScanTelemetry(
@@ -261,7 +266,15 @@ open class CodeWhispererClientAdaptorImpl(override val project: Project) : CodeW
                 it.timestamp(Instant.now())
             }
         }
+        requestBuilder.optOutPreference(getTelemetryOptoutPreference())
     }
+
+    private fun getTelemetryOptoutPreference() =
+        if (AwsSettings.getInstance().isTelemetryEnabled) {
+            OptOutPreference.OPTIN
+        } else {
+            OptOutPreference.OPTOUT
+        }
 
     override fun dispose() {
         if (this::mySigv4Client.isLazyInitialized) {
