+++ jetbrains-core/tst/software/aws/toolkits/jetbrains/services/codewhisperer/CodeWhispererTelemetryServiceTest.kt
@@ -4,8 +4,6 @@
 package software.aws.toolkits.jetbrains.services.codewhisperer
 
 import com.intellij.openapi.application.ApplicationManager
-import com.intellij.openapi.editor.VisualPosition
-import com.intellij.openapi.project.Project
 import com.intellij.testFramework.ApplicationRule
 import com.intellij.testFramework.DisposableRule
 import com.intellij.testFramework.ProjectRule
@@ -26,13 +24,9 @@ import org.mockito.kotlin.spy
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.verifyNoInteractions
 import org.mockito.kotlin.whenever
-import software.amazon.awssdk.services.codewhispererruntime.model.Completion
-import software.amazon.awssdk.services.codewhispererruntime.model.Import
-import software.amazon.awssdk.services.codewhispererruntime.model.Reference
 import software.aws.toolkits.core.telemetry.MetricEvent
 import software.aws.toolkits.core.telemetry.TelemetryBatcher
 import software.aws.toolkits.core.telemetry.TelemetryPublisher
-import software.aws.toolkits.core.utils.test.aString
 import software.aws.toolkits.jetbrains.AwsToolkit
 import software.aws.toolkits.jetbrains.core.credentials.ManagedBearerSsoConnection
 import software.aws.toolkits.jetbrains.core.credentials.ToolkitConnectionManager
@@ -41,36 +35,18 @@ import software.aws.toolkits.jetbrains.services.codewhisperer.credentials.CodeWh
 import software.aws.toolkits.jetbrains.services.codewhisperer.credentials.CodeWhispererLoginType
 import software.aws.toolkits.jetbrains.services.codewhisperer.credentials.MockCodeWhispererClientAdaptor
 import software.aws.toolkits.jetbrains.services.codewhisperer.explorer.CodeWhispererExplorerActionManager
-import software.aws.toolkits.jetbrains.services.codewhisperer.language.CodeWhispererProgrammingLanguage
-import software.aws.toolkits.jetbrains.services.codewhisperer.language.languages.CodeWhispererJava
-import software.aws.toolkits.jetbrains.services.codewhisperer.language.languages.CodeWhispererPython
-import software.aws.toolkits.jetbrains.services.codewhisperer.model.CaretContext
-import software.aws.toolkits.jetbrains.services.codewhisperer.model.CaretPosition
-import software.aws.toolkits.jetbrains.services.codewhisperer.model.Chunk
-import software.aws.toolkits.jetbrains.services.codewhisperer.model.DetailContext
-import software.aws.toolkits.jetbrains.services.codewhisperer.model.FileContextInfo
-import software.aws.toolkits.jetbrains.services.codewhisperer.model.LatencyContext
-import software.aws.toolkits.jetbrains.services.codewhisperer.model.RecommendationContext
-import software.aws.toolkits.jetbrains.services.codewhisperer.model.SessionContext
-import software.aws.toolkits.jetbrains.services.codewhisperer.model.SupplementalContextInfo
-import software.aws.toolkits.jetbrains.services.codewhisperer.model.TriggerTypeInfo
 import software.aws.toolkits.jetbrains.services.codewhisperer.service.CodeWhispererAutomatedTriggerType
 import software.aws.toolkits.jetbrains.services.codewhisperer.service.CodeWhispererInvocationStatus
 import software.aws.toolkits.jetbrains.services.codewhisperer.service.CodeWhispererUserGroup
 import software.aws.toolkits.jetbrains.services.codewhisperer.service.CodeWhispererUserGroupSettings
 import software.aws.toolkits.jetbrains.services.codewhisperer.service.CodeWhispererUserGroupStates
-import software.aws.toolkits.jetbrains.services.codewhisperer.service.RequestContext
-import software.aws.toolkits.jetbrains.services.codewhisperer.service.ResponseContext
 import software.aws.toolkits.jetbrains.services.codewhisperer.telemetry.CodeWhispererTelemetryService
-import software.aws.toolkits.jetbrains.services.codewhisperer.util.CrossFileStrategy
-import software.aws.toolkits.jetbrains.services.codewhisperer.util.UtgStrategy
 import software.aws.toolkits.jetbrains.services.telemetry.NoOpPublisher
 import software.aws.toolkits.jetbrains.services.telemetry.TelemetryService
 import software.aws.toolkits.jetbrains.settings.AwsSettings
 import software.aws.toolkits.telemetry.CodewhispererCompletionType
 import software.aws.toolkits.telemetry.CodewhispererPreviousSuggestionState
 import software.aws.toolkits.telemetry.CodewhispererSuggestionState
-import software.aws.toolkits.telemetry.CodewhispererTriggerType
 import java.time.Duration
 import kotlin.random.Random
 
@@ -589,183 +565,4 @@ class CodeWhispererTelemetryServiceTest {
 
         verifyNoInteractions(mockClient)
     }
-}
-
-// TODO: move these to testUtil
-fun aRequestContext(project: Project, myFileContextInfo: FileContextInfo? = null, mySupplementalContextInfo: SupplementalContextInfo? = null): RequestContext {
-    val triggerType = listOf(CodewhispererTriggerType.AutoTrigger, CodewhispererTriggerType.OnDemand).random()
-    val automatedTriggerType = if (triggerType == CodewhispererTriggerType.AutoTrigger) {
-        listOf(
-            CodeWhispererAutomatedTriggerType.IdleTime(),
-            CodeWhispererAutomatedTriggerType.Enter(),
-            CodeWhispererAutomatedTriggerType.SpecialChar('a'),
-            CodeWhispererAutomatedTriggerType.IntelliSense()
-        ).random()
-    } else {
-        CodeWhispererAutomatedTriggerType.Unknown()
-    }
-
-    return RequestContext(
-        project,
-        mock(),
-        TriggerTypeInfo(triggerType, automatedTriggerType),
-        CaretPosition(Random.nextInt(), Random.nextInt()),
-        fileContextInfo = myFileContextInfo ?: aFileContextInfo(),
-        supplementalContext = mySupplementalContextInfo ?: aSupplementalContextInfo(),
-        null,
-        LatencyContext(
-            Random.nextLong(),
-            Random.nextLong(),
-            Random.nextLong(),
-            Random.nextLong(),
-            Random.nextDouble(),
-            Random.nextLong(),
-            Random.nextLong(),
-            Random.nextLong(),
-            Random.nextLong(),
-            Random.nextLong(),
-            Random.nextLong(),
-            aString()
-        )
-    )
-}
-
-fun aSupplementalContextInfo(myContents: List<Chunk>? = null, myIsUtg: Boolean? = null, myLatency: Long? = null): SupplementalContextInfo {
-    val contents = mutableListOf<Chunk>()
-    val numberOfContent = Random.nextInt(1, 4)
-    repeat(numberOfContent) {
-        contents.add(
-            Chunk(
-                content = aString(),
-                path = aString(),
-            )
-        )
-    }
-
-    val isUtg = Random.nextBoolean()
-    val latency = Random.nextLong(from = 0L, until = 100L)
-
-    return SupplementalContextInfo(
-        isUtg = myIsUtg ?: isUtg,
-        latency = myLatency ?: latency,
-        contents = myContents ?: contents,
-        targetFileName = aString(),
-        strategy = if (myIsUtg ?: isUtg) UtgStrategy.ByName else CrossFileStrategy.OpenTabsBM25
-    )
-}
-
-fun aRecommendationContext(): RecommendationContext {
-    val details = mutableListOf<DetailContext>()
-    val size = Random.nextInt(1, 5)
-    for (i in 1..size) {
-        details.add(
-            DetailContext(
-                aString(),
-                aCompletion(),
-                aCompletion(),
-                listOf(true, false).random(),
-                listOf(true, false).random(),
-                aString(),
-                CodewhispererCompletionType.Line
-            )
-        )
-    }
-
-    return RecommendationContext(
-        details,
-        aString(),
-        aString(),
-        VisualPosition(Random.nextInt(1, 100), Random.nextInt(1, 100))
-    )
-}
-
-/**
- * util to generate a RecommendationContext and a SessionContext given expected decisions
- */
-fun aRecommendationContextAndSessionContext(decisions: List<CodewhispererSuggestionState>): Pair<RecommendationContext, SessionContext> {
-    val table = CodewhispererSuggestionState.values().associateWith { 0 }.toMutableMap()
-    decisions.forEach {
-        table[it]?.let { curCount -> table[it] = 1 + curCount }
-    }
-
-    val details = mutableListOf<DetailContext>()
-    decisions.forEach { decision ->
-        val toAdd = if (decision == CodewhispererSuggestionState.Empty) {
-            val completion = aCompletion("", true, 0, 0)
-            DetailContext(aString(), completion, completion, Random.nextBoolean(), Random.nextBoolean(), aString(), CodewhispererCompletionType.Unknown)
-        } else if (decision == CodewhispererSuggestionState.Discard) {
-            val completion = aCompletion()
-            DetailContext(aString(), completion, completion, true, Random.nextBoolean(), aString(), CodewhispererCompletionType.Unknown)
-        } else {
-            val completion = aCompletion()
-            DetailContext(aString(), completion, completion, false, Random.nextBoolean(), aString(), CodewhispererCompletionType.Unknown)
-        }
-
-        details.add(toAdd)
-    }
-
-    val recommendationContext = RecommendationContext(
-        details,
-        aString(),
-        aString(),
-        VisualPosition(Random.nextInt(1, 100), Random.nextInt(1, 100))
-    )
-
-    val selectedIndex = decisions.indexOfFirst { it == CodewhispererSuggestionState.Accept }.let {
-        if (it != -1) {
-            it
-        } else {
-            0
-        }
-    }
-
-    val seen = mutableSetOf<Int>()
-    decisions.forEachIndexed { index, decision ->
-        if (decision != CodewhispererSuggestionState.Unseen) {
-            seen.add(index)
-        }
-    }
-
-    val sessionContext = SessionContext(
-        selectedIndex = selectedIndex,
-        seen = seen
-    )
-    return recommendationContext to sessionContext
-}
-
-fun aCompletion(content: String? = null, isEmpty: Boolean = false, referenceCount: Int? = null, importCount: Int? = null): Completion {
-    val myReferenceCount = referenceCount ?: Random.nextInt(0, 4)
-    val myImportCount = importCount ?: Random.nextInt(0, 4)
-
-    val references = List(myReferenceCount) {
-        Reference.builder()
-            .licenseName(aString())
-            .build()
-    }
-
-    val imports = List(myImportCount) {
-        Import.builder()
-            .statement(aString())
-            .build()
-    }
-
-    return Completion.builder()
-        .content(content ?: if (!isEmpty) aString() else "")
-        .references(references)
-        .mostRelevantMissingImports(imports)
-        .build()
-}
-
-fun aResponseContext(): ResponseContext = ResponseContext(aString())
-
-fun aFileContextInfo(language: CodeWhispererProgrammingLanguage? = null): FileContextInfo {
-    val caretContextInfo = CaretContext(aString(), aString(), aString())
-    val fileName = aString()
-
-    val programmingLanguage = language ?: listOf(
-        CodeWhispererPython.INSTANCE,
-        CodeWhispererJava.INSTANCE
-    ).random()
-
-    return FileContextInfo(caretContextInfo, fileName, programmingLanguage)
 }
